requirements.txt
tensorflow-macos
numpy
matplotlib